feat: Refactor user to member and update UI
This commit refactors the agora_dao.cairo smart contract to use the term "member" instead of "user" for clarity. This includes renaming user_counter to member_counter and the users map to members.

The frontend has been updated to reflect these changes in the contract ABI and related components. Additionally, the landing page has been redesigned, and a new dialog for adding roles has been introduced.